fix(server.ts): change the case of the port variable from lowercase to uppercase (PORT) feature(server.ts): add support for the process.env.PORT environment variable
The port variable is now named PORT, which improves consistency with naming conventions as PORT is a constant. The support for an environment variable allows the application to be more flexible as it can now run on any available port specified via the process.env.PORT environment variable.

fix(AsaComplement.js, InSummary.js): add ProductStyle component

The ProductStyle component is imported and used in the AsaComplement and InSummary components. This component is responsible for styling the products in a specific way.

fix(apportMineraux.js): fix id value

The id value in the apportMineraux.js file has been corrected from "AMiS2" to "AMiS0" to match the correct product.

fix(apresShampooing.js): fix typo in text

A typo in the text of the apresShampooing.js file has been fixed. The word "rédeuisant" has been corrected to "réduisant" for proper spelling.

fix(detox.js): update colors and text

The colors and text in the detox.js file have been updated. New color values have been added for bgColorInSummary0, bgColorInSummary1, linearGradient0, linearGradient1, and shadowColor. The text in the inSummary section has been modified to include bullet points and headings for better readability.

fix(regenerateurCellulaire.js): change the IDs of two items in the regenerateurCellulaire data to match the correct IDs (RCiSB1 and RCiSBE2)
fix(ProductPage.js): add a new CSS class "blocTitleImg" to wrap the title and image elements in the product details section
fix(_flex.scss): add the new CSS class "blocTitleImg" to the list of flex classes
fix(_width.scss): add a new width rule for the "blocTitleImg" class to make it 80% of the viewport width
fix(_productCard.scss): comment out the height, margin-top, and padding-left rules for the "bgProductDetails" class

The IDs of two items in the regenerateurCellulaire data were incorrect and have been fixed. The "blocTitleImg" class was added to properly style the title and image elements in the product details section. The necessary CSS classes were updated to include the new "blocTitleImg" class. The width of the "blocTitleImg" class was adjusted to be 80% of the viewport width. The height, margin-top, and padding-left rules for the "bgProductDetails" class were commented out. These changes improve the consistency and styling of the product page.