chore(server.ts): change port variable name from 'port' to 'PORT' feature(server.ts): add support for process.env.PORT environment variable The port variable is now named PORT, which improves consistency with naming conventions as PORT is a constant. The support for an environment variable allows the application to be more flexible as it can now run on any available port specified via the process.env.PORT environment variable.
feat(privacy.js, terms.js): update legal titles and content
The legal titles and content in the privacy.js and terms.js files have been updated to reflect the correct titles and information. This ensures that the privacy policy and terms of use are accurate and up to date.

style(_margin.scss): adjust margin-bottom for legalContainer ul and legalTitle
The margin-bottom for the legalContainer ul and legalTitle elements has been adjusted to 20px. This improves the spacing and layout of the elements.

style(_padding.scss): adjust padding for legalTitle
The padding for the legalTitle element has been adjusted to 90px. This improves the spacing and layout of the element.